Refactor ioc.config.ts to avoid circular refs
A minimal set of files should include ioc.config.ts or else errors such
as the following occur.

>nyc mocha test/lib/util/Config.test.ts
Error: Cannot register a type name as a singleton without a "to" token to=undefined from=Config

The reason for this is that a file imports Config.ts, but Config.ts
imports ioc.config.ts.

TODO: Consider renaming Controller to ScannerFactory.
