+++ test/test-related-lib/TestOverrides.ts
@@ -1,7 +1,9 @@
 import { EnvOverridable } from "../../src/Constants";
 import os = require('os');
 import path = require('path');
-import { Controller, Services } from "../../src/ioc.config";
+import { Services } from "../../src/Constants";
+import { Controller } from "../../src/Controller";
+import { registerAll } from "../../src/ioc.config";
 
 export class TestOverrides implements EnvOverridable {
 	public getSfdxScannerPath(): string {
@@ -18,5 +20,5 @@ function setupTestAlternatives(): void {
 export function initializeTestSetup(): void {
 	container.reset();
 	setupTestAlternatives();
-	Controller.registerAll();
+	registerAll();
 }
